+++ packages/react-workspace/spright-react/src/icons/breakpoint-hit-disabled/index.ts
@@ -1,0 +1,8 @@
+'use client';
+
+import { IconBreakpointHitDisabled, iconBreakpointHitDisabledTag } from '@ni/spright-components/dist/esm/icons/breakpoint-hit-disabled';
+import { wrap } from '../../utilities/react-wrapper';
+
+export type { IconBreakpointHitDisabled };
+export { iconBreakpointHitDisabledTag };
+export const SprightIconBreakpointHitDisabled = wrap(IconBreakpointHitDisabled);