+++ packages/react-workspace/spright-react/src/icons/breakpoint-hit/index.ts
@@ -1,0 +1,8 @@
+'use client';
+
+import { IconBreakpointHit, iconBreakpointHitTag } from '@ni/spright-components/dist/esm/icons/breakpoint-hit';
+import { wrap } from '../../utilities/react-wrapper';
+
+export type { IconBreakpointHit };
+export { iconBreakpointHitTag };
+export const SprightIconBreakpointHit = wrap(IconBreakpointHit);