+++ packages/spright-components/src/icon-base/tests/icon-metadata.ts
@@ -10,17 +10,38 @@ export const iconMetadata: {
     readonly [key in IconName]: IconMetadata;
 } = {
     /* eslint-disable @typescript-eslint/naming-convention */
+    IconBreakpointConditional: {
+        tags: ['debug']
+    },
+    IconBreakpointDisabled: {
+        tags: ['debug']
+    },
+    IconBreakpointEnabled: {
+        tags: ['debug']
+    },
+    IconBreakpointExecutionPointer: {
+        tags: ['debug']
+    },
+    IconBreakpointHit: {
+        tags: ['debug']
+    },
+    IconBreakpointHitDisabled: {
+        tags: ['debug']
+    },
+    IconBreakpointHover: {
+        tags: ['debug']
+    },
     IconNigelChat: {
         tags: ['nigel', 'chat', 'ai']
     },
     IconWorkItemCalendarWeek: {
         tags: ['reservation', 'booking', 'schedule', 'time slot']
     },
-    IconWorkItemForklift: {
-        tags: ['logistics', 'delivery', 'transport', 'shipping']
-    },
     IconWorkItemCalipers: {
         tags: ['calibration', 'quality', 'measurement', 'standards']
+    },
+    IconWorkItemForklift: {
+        tags: ['logistics', 'delivery', 'transport', 'shipping']
     },
     IconWorkItemRectangleCheckLines: {
         tags: ['testing', 'verification', 'quality assurance', 'checklist']