+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Mon, 04 May 2026 17:04:06 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 27 May 2026 16:39:52 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 35.9.3
+
+Wed, 27 May 2026 16:39:52 GMT
+
+### Patches
+
+- Bump @ni/unit-format to v1.0.5
+
+## 35.9.2
+
+Tue, 26 May 2026 18:42:52 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v8.18.1
+
+## 35.9.1
+
+Tue, 19 May 2026 17:01:52 GMT
+
+### Patches
+
+- Add color-scheme to theme provider for dark scrollbars ([ni/nimble@3561c88](https://github.com/ni/nimble/commit/3561c88b0bb55fcbaf9345d2915f0e8ac310b5fc))
+
+## 35.9.0
+
+Tue, 19 May 2026 16:29:54 GMT
+
+### Minor changes
+
+- Add appearance-indeterminate property for checkbox ([ni/nimble@38b5b13](https://github.com/ni/nimble/commit/38b5b13eceb1f7c1da8342dc583d79769e69ca7a))
 
 ## 35.8.0